Fix Telegram ID overflow: Migrated user_id to BigInteger and added auto-migration logic

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime
 from sqlalchemy.sql import func
 from app.database import Base
 
@@ -6,7 +6,7 @@ class User(Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True)
-    user_id = Column(Integer, unique=True, index=True)
+    user_id = Column(BigInteger, unique=True, index=True)
     username = Column(String, nullable=True)
     full_name = Column(String, nullable=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
@@ -15,7 +15,7 @@ class QueryLog(Base):
     __tablename__ = "query_logs"
 
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, index=True)
+    user_id = Column(BigInteger, index=True)
     username_or_url = Column(String, index=True)
     status = Column(String)  # success, error, not_found
     created_at = Column(DateTime(timezone=True), server_default=func.now())
